fttp_cost/market_engine.py
from tavily import TavilyClient
import os
from dotenv import load_dotenv
import re

load_dotenv()

client = TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))

def extract_price(text):
    numbers = re.findall(r'\d+', text)
    if numbers:
        return int(numbers[0])
    return None

def get_market_rates():

    try:
        queries = [
            "FTTH trenching cost per km India latest",
            "fiber optic cable cost per km India telecom",
            "telecom labour cost percentage India FTTH"
        ]

        results = []

        for q in queries:
            res = client.search(query=q, max_results=2)
            results.append(str(res))

        combined_text = " ".join(results)

        # --- Extract approximate values ---
        civil = extract_price(combined_text) or 420000
        fibre = 80000   # fallback (stable industry avg)
        labour_ratio = 0.28

        # --- Normalize unrealistic values ---
        if civil < 100000:
            civil = 400000
        if civil > 1000000:
            civil = 500000

        return {
            "civil_per_km": civil,
            "fibre_per_km": fibre,
            "labour_ratio": labour_ratio
        }

    except Exception as e:
        # fallback (SAFE)
        return {
            "civil_per_km": 420000,
            "fibre_per_km": 80000,
            "labour_ratio": 0.28
        }