fttp_cost/requirements.txt
streamlit==1.37.0

langgraph==0.2.16
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.12

langchain-groq==0.1.9
groq==0.9.0

tavily-python==0.3.5

osmnx==2.0.0
geopandas==1.0.1
shapely==2.0.4
networkx==3.3

geopy==2.4.1

chromadb==0.5.0

python-dotenv==1.0.1

pandas==2.2.2
numpy==1.26.4
requests==2.32.3
pip install folium